Move exiv2 app to a separated folder
This change revealed that i18n.h needs to be a public header of
exiv2lib.
