+++ exiv2app/getopt_win32.h
