+++ exiv2app/utils.hpp
