+++ exiv2lib/canonmn_int.cpp
@@ -29,11 +29,11 @@
 #include <exiv2/types.hpp>
 #include <exiv2/value.hpp>
 #include <exiv2/exif.hpp>
+#include <exiv2/i18n.h>
 
 #include "makernote_int.hpp"
 #include "canonmn_int.hpp"
 #include "tags_int.hpp"
-#include "i18n.h"                // NLS support.
 
 // + standard includes
 #include <string>
