+++ exiv2lib/include/exiv2/i18n.h
