+++ exiv2lib/minoltamn_int.cpp
@@ -31,11 +31,11 @@
 #include <exiv2/value.hpp>
 #include <exiv2/exif.hpp>
 #include <exiv2/datasets.hpp>
+#include <exiv2/i18n.h>
 
 #include "minoltamn_int.hpp"
 #include "tags_int.hpp"
 #include "makernote_int.hpp"
-#include "i18n.h"                // NLS support.
 
 #include <string>
 #include <sstream>
