+++ exiv2lib/olympusmn_int.cpp
@@ -33,11 +33,11 @@
 #include <exiv2/types.hpp>
 #include <exiv2/value.hpp>
 #include <exiv2/image.hpp>
+#include <exiv2/i18n.h>
 
 #include "olympusmn_int.hpp"
 #include "tags_int.hpp"
 #include "makernote_int.hpp"
-#include "i18n.h"                // NLS support.
 
 // + standard includes
 #include <string>
