+++ exiv2lib/sigmamn_int.cpp
@@ -30,10 +30,10 @@
 // included header files
 #include <exiv2/types.hpp>
 #include <exiv2/value.hpp>
+#include <exiv2/i18n.h>
 
 #include "sigmamn_int.hpp"
 #include "tags_int.hpp"
-#include "i18n.h"                // NLS support.
 
 // + standard includes
 #include <string>
