+++ exiv2lib/tiffimage_int.cpp
@@ -1,8 +1,9 @@
+#include <exiv2/i18n.h>
+
 #include "tiffimage_int.hpp"
 
 #include "makernote_int.hpp"
 #include "tiffvisitor_int.hpp"
-#include "i18n.h"                // NLS support.
 
 // Shortcuts for the newTiffBinaryArray templates.
 #define EXV_BINARY_ARRAY(arrayCfg, arrayDef) (newTiffBinaryArray0<&arrayCfg, EXV_COUNTOF(arrayDef), arrayDef>)
